# etcdctl: improve `etcdctl cluster-health`