--- a/etcdctl/command/cluster_health.go
+++ b/etcdctl/command/cluster_health.go
@@ -21,6 +21,7 @@ import (
 	"net/http"
 	"os"
 	"os/signal"
+	"sync"
 	"time"
 
 	"github.com/coreos/etcd/Godeps/_workspace/src/github.com/codegangsta/cli"
@@ -70,51 +71,60 @@ func handleClusterHealth(c *cli.Context) {
 
 	for {
 		health := false
+		wait := sync.WaitGroup{}
+		wait.Add(len(ms))
 		for _, m := range ms {
-			if len(m.ClientURLs) == 0 {
-				fmt.Printf("member %s is unreachable: no available published client urls\n", m.ID)
-				continue
-			}
+			go func(m client.Member) {
+				defer wait.Done()
 
-			checked := false
-			for _, url := range m.ClientURLs {
-				resp, err := hc.Get(url + "/health")
-				if err != nil {
-					fmt.Printf("failed to check the health of member %s on %s: %v\n", m.ID, url, err)
-					continue
+				if len(m.ClientURLs) == 0 {
+					fmt.Printf("member %s is unreachable: no available published client urls\n", m.ID)
+					return
 				}
 
-				result := struct{ Health string }{}
-				nresult := struct{ Health bool }{}
-				bytes, err := ioutil.ReadAll(resp.Body)
-				if err != nil {
-					fmt.Printf("failed to check the health of member %s on %s: %v\n", m.ID, url, err)
-					continue
-				}
-				resp.Body.Close()
+				checked := false
+				for _, url := range m.ClientURLs {
+					resp, err := hc.Get(url + "/health")
+					if err != nil {
+						fmt.Printf("failed to check the health of member %s on %s: %v\n", m.ID, url, err)
+						continue
+					}
 
-				err = json.Unmarshal(bytes, &result)
-				if err != nil {
-					err = json.Unmarshal(bytes, &nresult)
-				}
-				if err != nil {
-					fmt.Printf("failed to check the health of member %s on %s: %v\n", m.ID, url, err)
-					continue
-				}
+					result := struct{ Health string }{}
+					nresult := struct{ Health bool }{}
+					bytes, err := ioutil.ReadAll(resp.Body)
+					if err != nil {
+						fmt.Printf("failed to check the health of member %s on %s: %v\n", m.ID, url, err)
+						continue
+					}
+					resp.Body.Close()
 
-				checked = true
-				if result.Health == "true" || nresult.Health == true {
-					health = true
-					fmt.Printf("member %s is healthy: got healthy result from %s\n", m.ID, url)
-				} else {
-					fmt.Printf("member %s is unhealthy: got unhealthy result from %s\n", m.ID, url)
+					err = json.Unmarshal(bytes, &result)
+					if err != nil {
+						err = json.Unmarshal(bytes, &nresult)
+					}
+					if err != nil {
+						fmt.Printf("failed to check the health of member %s on %s: %v\n", m.ID, url, err)
+						continue
+					}
+
+					checked = true
+					if result.Health == "true" || nresult.Health == true {
+						health = true
+						fmt.Printf("member %s is healthy: got healthy result from %s\n", m.ID, url)
+					} else {
+						fmt.Printf("member %s is unhealthy: got unhealthy result from %s\n", m.ID, url)
+					}
+					break
 				}
-				break
-			}
-			if !checked {
-				fmt.Printf("member %s is unreachable: %v are all unreachable\n", m.ID, m.ClientURLs)
-			}
+				if !checked {
+					fmt.Printf("member %s is unreachable: %v are all unreachable\n", m.ID, m.ClientURLs)
+				}
+			}(m)
 		}
+
+		wait.Wait()
+
 		if health {
 			fmt.Println("cluster is healthy")
 		} else {