Usually, there are 3 or 5 members in an etcd cluster, `etcdctl cluster-health` will check those members' health sequentially. Although 3 and 5 is not a big number, every member will sleep 250ms to 750ms to ensure raft algorithm making progress on it, so that it can determine if itself healthy. In addition some network latency(or request timeout, which is much more terrible), user must wait about at least 3 to 5 seconds before they finally get the cluster health information.

I think it's a better choice to check members health status concurrently，especially considered network latency and request time out.