@jonboulle 2015-10-20
Actually, on looking more closely at the existing logic, this isn't right; currently we just return as soon as the first member successfully reports healthy or unhealthy. @mqliang are you having multiple members time out before they actually manage to report health?