@mqliang 2015-10-20
@jonboulle  Great thanks for your review!

Assume we have a 5 member cluster, A B C D E. When we run `etcdctl cluster-health`, original logic will check those members' health sequentially. If every machine is healthy and work fine, but unfortunately, request to machine A, B will time out for some network reasons, after retry through `all` the client-url of A and B, we print "A and B are unreachable". Request to machine C and D and E return immediately and indicate they are healthy. In such a case, we spend too much time to wait for the response from A and B. 

Usually, when we do some I/O operations, such as network request or disk operation, which may block and would not return immediately, we need process asynchronously or concurrently.

If my concurrent way looks good to you, I am glad to rework this PR to make the code more readable.